fix(knowledge): address review findings for PR #1727 (round 2)
Apply swarm-pr-feedback fixes for review findings F-001..F-006:

* F-001: Add PRR-015 re-archive guard to knowledge_archive tool
  (src/tools/knowledge-archive.ts:208-218) preventing
  archived_from corruption on duplicate archive calls.

* F-002: Rewrite docs/releases/pending/1716-knowledge-lifecycle-consistency.md
  to remove the false 'Out of scope' bullet about three filter sites that
  were updated, expand G4 scope to all 10 isActiveStatus() consumer sites,
  and add concrete test-file references for every acceptance criterion.

* F-003: Split 3 over-cap test files (FR-006 500-line cap) into 11 focused
  files with shared helpers and barrel re-exports:
  - knowledge-archive.test.ts -> tool-{swarm,hive,post-archive-hook}.test.ts
  - knowledge-config.test.ts -> schema-{defaults,validation}.test.ts
  - knowledge.test.ts -> tool-{quarantine,restore,list,migrate}.test.ts

* F-004: Fix afterEach async cleanup leak in knowledge-floor-action.test.ts
  by switching to static rmSync import + realpathSync(mkdtempSync) pattern.

* F-005: Fix mock isolation patterns:
  - knowledge-reader.test.ts: spread realStore exports in mock.module
  - knowledge commands mocks: targeted mockReset() instead of global
    mock.restore()

* F-006: Close behavioral coverage gaps:
  - floorAction:'quarantine' mode test added to knowledge-floor-action.test.ts
  - TTL-unarchive recent_negative_phase_count + last_demotion_phase reset
    assertions added to knowledge-unarchive.test.ts

Branch state:
- Merged origin/main (v7.107.3) to resolve CONFLICTING/DIRTY state
- 4 release-management files auto-merged to main's version (release-please owned)
- 2 content conflicts in src/hooks/knowledge-reader.ts and
  tests/unit/commands/knowledge.test.ts resolved by taking PR side
  (preserves isActiveStatus helper and mockUnarchiveEntry setup)

Verification:
- bun run typecheck: exit 0
- bunx biome ci: exit 0
- bun run drift:check: clean
- 131/131 individual test files pass (no regressions)
- Pre-existing failures preserved (8 fail, all documented in PR body)

NOT included in this commit:
- F-007 (writePhaseEvidence dead export): DISPROVED by verification; function
  is still actively used by runner.ts:27 and downstream callers. Removed
  attempted edit.
