+++ tests/unit/config/_knowledge-config-helpers.ts
@@ -1,0 +1,99 @@
+import type { KnowledgeConfig } from '../../../src/config/schema.js';
+
+/**
+ * Full config object used in the "should succeed when parsing a full config with
+ * all fields set" test. Note: G7 promoted_demotion fields are NOT included
+ * because the original test predates those schema additions.
+ */
+export const FULL_CONFIG_OBJECT = {
+	enabled: false,
+	swarm_max_entries: 500,
+	hive_max_entries: 1000,
+	auto_promote_days: 30,
+	max_inject_count: 10,
+	delegate_max_inject_count: 6,
+	inject_char_budget: 3_000,
+	max_lesson_display_chars: 200,
+	dedup_threshold: 0.8,
+	scope_filter: ['global', 'project'] as const,
+	hive_enabled: false,
+	rejected_max_entries: 50,
+	validation_enabled: false,
+	evergreen_confidence: 0.95,
+	evergreen_utility: 0.85,
+	low_utility_threshold: 0.25,
+	min_retrievals_for_utility: 5,
+	schema_version: 2,
+	same_project_weight: 1.0,
+	cross_project_weight: 0.5,
+	min_encounter_score: 0.1,
+	initial_encounter_score: 1.0,
+	encounter_increment: 0.1,
+	max_encounter_score: 10.0,
+	default_max_phases: 10,
+	todo_max_phases: 3,
+	sweep_enabled: true,
+	// G2/G3 (#1715) feedback-loop closure fields
+	confidence_floor_action: 'quarantine' as const,
+	confidence_floor_min_outcomes: 5,
+	confidence_floor_signal_threshold: -0.1,
+	contradiction_threshold_action: 'tag_only' as const,
+	contradiction_quarantine_threshold: 5,
+	contradiction_quarantine_window_days: 14,
+	realtime_learning_nudge: {
+		enabled: false,
+		first_after_tool_calls: 5,
+		repeat_after_tool_calls: 15,
+	},
+	directive_min_confidence: 0.75,
+	enrichment: {
+		max_calls_per_day: 30,
+		quota_window: 'utc' as const,
+	},
+};
+
+export const MINIMAL_KNOWLEDGE_CONFIG: KnowledgeConfig = {
+	enabled: true,
+	swarm_max_entries: 100,
+	hive_max_entries: 200,
+	auto_promote_days: 90,
+	max_inject_count: 5,
+	delegate_max_inject_count: 8,
+	inject_char_budget: 2_000,
+	max_lesson_display_chars: 120,
+	dedup_threshold: 0.6,
+	scope_filter: ['global'],
+	hive_enabled: true,
+	rejected_max_entries: 20,
+	validation_enabled: true,
+	evergreen_confidence: 0.9,
+	evergreen_utility: 0.8,
+	low_utility_threshold: 0.3,
+	min_retrievals_for_utility: 3,
+	schema_version: 1,
+	same_project_weight: 1.0,
+	cross_project_weight: 0.5,
+	min_encounter_score: 0.1,
+	initial_encounter_score: 1.0,
+	encounter_increment: 0.1,
+	max_encounter_score: 10.0,
+	default_max_phases: 10,
+	todo_max_phases: 3,
+	sweep_enabled: true,
+	confidence_floor_action: 'demote',
+	confidence_floor_min_outcomes: 3,
+	confidence_floor_signal_threshold: 0,
+	contradiction_threshold_action: 'quarantine',
+	contradiction_quarantine_threshold: 3,
+	contradiction_quarantine_window_days: 30,
+	realtime_learning_nudge: {
+		enabled: true,
+		first_after_tool_calls: 10,
+		repeat_after_tool_calls: 25,
+	},
+	directive_min_confidence: 0.75,
+	enrichment: {
+		max_calls_per_day: 30,
+		quota_window: 'utc',
+	},
+};
