+++ tests/unit/tools/knowledge-archive.test.ts
@@ -1,609 +1,14 @@
-import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
-import { existsSync, mkdirSync, rmSync, writeFileSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
-import {
-	type ArchivedEvent,
-	readHiveKnowledgeEvents,
-	readKnowledgeEvents,
-} from '../../../src/hooks/knowledge-events';
-import {
-	appendKnowledge,
-	getArchivedKnowledgeIds,
-	readKnowledge,
-	resolveHiveKnowledgePath,
-	resolveSwarmKnowledgePath,
-} from '../../../src/hooks/knowledge-store';
-import type {
-	HiveKnowledgeEntry,
-	SwarmKnowledgeEntry,
-} from '../../../src/hooks/knowledge-types';
-import { knowledge_archive } from '../../../src/tools/knowledge-archive';
-
-function makeSwarmEntry(id: string): SwarmKnowledgeEntry {
-	return {
-		id,
-		tier: 'swarm',
-		lesson: `Lesson ${id} with enough characters to be valid`,
-		category: 'process',
-		tags: [],
-		scope: 'global',
-		confidence: 0.5,
-		status: 'candidate',
-		confirmed_by: [],
-		project_name: 'test',
-		retrieval_outcomes: {
-			applied_count: 0,
-			succeeded_after_count: 0,
-			failed_after_count: 0,
-		},
-		schema_version: 2,
-		created_at: new Date().toISOString(),
-		updated_at: new Date().toISOString(),
-	};
-}
-
-function makeHiveEntry(id: string): HiveKnowledgeEntry {
-	return {
-		id,
-		tier: 'hive',
-		lesson: `Hive Lesson ${id} with enough characters to be valid`,
-		category: 'architecture',
-		tags: [],
-		scope: 'global',
-		confidence: 0.8,
-		status: 'established',
-		confirmed_by: [],
-		source_project: 'original-project',
-		encounter_score: 1.5,
-		retrieval_outcomes: {
-			applied_count: 5,
-			succeeded_after_count: 3,
-			failed_after_count: 0,
-		},
-		schema_version: 2,
-		created_at: new Date().toISOString(),
-		updated_at: new Date().toISOString(),
-	};
-}
-
-const ctx = (directory: string): any => ({
-	directory,
-	sessionID: 'sess-1',
-	agent: 'architect',
-});
-
-describe('knowledge_archive', () => {
-	let dir: string;
-	let swarmPath: string;
-	let hivePath: string;
-	let previousHome: string | undefined;
-	let previousXdgDataHome: string | undefined;
-	let previousLocalAppData: string | undefined;
-
-	beforeEach(async () => {
-		previousHome = process.env.HOME;
-		previousXdgDataHome = process.env.XDG_DATA_HOME;
-		previousLocalAppData = process.env.LOCALAPPDATA;
-		dir = join(
-			tmpdir(),
-			`swarm-archive-${Date.now()}-${Math.random().toString(36).slice(2)}`,
-		);
-		mkdirSync(dir, { recursive: true });
-		// Redirect hive path resolution for all platforms:
-		// macOS (darwin) uses HOME; Linux uses XDG_DATA_HOME; Windows uses LOCALAPPDATA.
-		// resolveHiveKnowledgePath() reads process.env.HOME live per call (see knowledge-store.ts).
-		process.env.HOME = dir;
-		process.env.XDG_DATA_HOME = join(dir, 'xdg-data');
-		process.env.LOCALAPPDATA = join(dir, 'localappdata');
-		swarmPath = resolveSwarmKnowledgePath(dir);
-		hivePath = resolveHiveKnowledgePath();
-		await appendKnowledge(swarmPath, makeSwarmEntry('k1'));
-	});
-
-	afterEach(() => {
-		if (previousHome === undefined) delete process.env.HOME;
-		else process.env.HOME = previousHome;
-		if (previousXdgDataHome === undefined) delete process.env.XDG_DATA_HOME;
-		else process.env.XDG_DATA_HOME = previousXdgDataHome;
-		if (previousLocalAppData === undefined) delete process.env.LOCALAPPDATA;
-		else process.env.LOCALAPPDATA = previousLocalAppData;
-		rmSync(dir, { recursive: true, force: true });
-	});
-
-	describe('swarm-tier (default)', () => {
-		it('archives by default: sets status archived and keeps the entry', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'k1', reason: 'stale' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(true);
-			expect(parsed.mode).toBe('archive');
-			expect(parsed.tier).toBe('swarm');
-			expect(parsed.previous_status).toBe('candidate');
-			expect(parsed.status).toBe('archived');
-
-			const entries = await readKnowledge<SwarmKnowledgeEntry>(swarmPath);
-			expect(entries).toHaveLength(1);
-			expect(entries[0].status).toBe('archived');
-			// PRR-002: producer-side archived_from assertion — verify the tool
-			// records the prior status so unarchiveEntry can restore it.
-			expect(entries[0].archived_from).toBe('candidate');
-			expect(entries[0].archived_at).toBeTruthy();
-
-			const tomb = (await readKnowledgeEvents(dir)).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb).toHaveLength(1);
-			expect(tomb[0].entry_id).toBe('k1');
-			expect(tomb[0].actor).toBe('architect');
-			expect(tomb[0].reason).toBe('stale');
-			expect(tomb[0].previous_status).toBe('candidate');
-			expect(tomb[0].mode).toBe('archive');
-			expect(tomb[0].tier).toBe('swarm');
-		});
-
-		it('quarantines when mode=quarantine (routes through quarantineEntry — G5 #1716)', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'k1', reason: 'suspect', mode: 'quarantine', evidence: 'flaky' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(true);
-			expect(parsed.status).toBe('quarantined');
-			expect(parsed.tier).toBe('swarm');
-
-			// G5: the entry is MOVED to knowledge-quarantined.jsonl (not left
-			// in place with a flipped status). The main store no longer has it.
-			const swarmEntries = await readKnowledge<SwarmKnowledgeEntry>(swarmPath);
-			expect(swarmEntries).toHaveLength(0);
-
-			const quarantinePath = join(
-				resolveSwarmKnowledgePath(dir).replace(/knowledge\.jsonl$/, ''),
-				'knowledge-quarantined.jsonl',
-			);
-			const quarantined = await readKnowledge<
-				SwarmKnowledgeEntry & {
-					original_status: string;
-					quarantine_reason: string;
-				}
-			>(quarantinePath);
-			expect(quarantined).toHaveLength(1);
-			expect(quarantined[0].id).toBe('k1');
-			expect(quarantined[0].status).toBe('quarantined');
-			expect(quarantined[0].original_status).toBe('candidate');
-			expect(quarantined[0].quarantine_reason).toBe('suspect');
-
-			// G5: NO events tombstone is written for quarantine mode — the
-			// short-circuit returns before the tombstone block, and
-			// `quarantineEntry` writes its own audit trail (rejected-file
-			// tombstone) which is the canonical record for quarantine.
-			const tomb = (await readKnowledgeEvents(dir)).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb).toHaveLength(0);
-
-			// G5 (B5): the `queueMicrotask` skill-invalidation must NOT fire for
-			// quarantine mode. The only writer of `skill-stale-batch` is that
-			// microtask; if it (incorrectly) ran for quarantine, this event
-			// would appear. Skill-retirement is irreversible and is for
-			// archive/purge only.
-			// The microtask is queued before the test reads events; drain it by
-			// awaiting a microtask tick.
-			await new Promise((resolve) => queueMicrotask(resolve));
-			const skillEvents = (await readKnowledgeEvents(dir)).filter(
-				(e) => (e as { type?: string }).type === 'skill-stale-batch',
-			);
-			expect(skillEvents).toHaveLength(0);
-		});
-
-		it('refuses to purge without the admin flag', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'k1', reason: 'gone', mode: 'purge' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(false);
-			expect(parsed.error).toContain('allow_purge');
-			// Entry untouched.
-			const entries = await readKnowledge<SwarmKnowledgeEntry>(swarmPath);
-			expect(entries).toHaveLength(1);
-			expect(entries[0].status).toBe('candidate');
-		});
-
-		it('PRR-003: quarantine returns not-found (not false-success) for a missing id', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'missing', reason: 'suspect', mode: 'quarantine' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			// PRR-003: quarantineEntry silently returns void on not-found, so
-			// the tool MUST pre-check and report not-found rather than
-			// reporting success:true (matching archive mode's behavior).
-			expect(parsed.success).toBe(false);
-			expect(parsed.message).toBe('entry not found');
-		});
-
-		it('purges (hard-deletes) with allow_purge:true and still writes a tombstone', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'k1', reason: 'gone', mode: 'purge', allow_purge: true },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(true);
-			expect(parsed.status).toBe('purged');
-			expect(parsed.tier).toBe('swarm');
-			const entries = await readKnowledge<SwarmKnowledgeEntry>(swarmPath);
-			expect(entries).toHaveLength(0);
-			const tomb = (await readKnowledgeEvents(dir)).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb).toHaveLength(1);
-			expect(tomb[0].mode).toBe('purge');
-			expect(tomb[0].tier).toBe('swarm');
-		});
-
-		it('returns not found for an unknown id and writes no tombstone', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'missing', reason: 'x' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(false);
-			expect(parsed.message).toBe('entry not found');
-			expect(await readKnowledgeEvents(dir)).toHaveLength(0);
-		});
-
-		it('requires id and reason', async () => {
-			const noId = JSON.parse(
-				await knowledge_archive.execute({ reason: 'x' } as never, ctx(dir)),
-			);
-			expect(noId.success).toBe(false);
-			const noReason = JSON.parse(
-				await knowledge_archive.execute({ id: 'k1' } as never, ctx(dir)),
-			);
-			expect(noReason.success).toBe(false);
-		});
-	});
-
-	describe('hive-tier', () => {
-		beforeEach(async () => {
-			await appendKnowledge(hivePath, makeHiveEntry('hive-1'));
-		});
-
-		it('archives hive entry when tier=hive', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'hive-1', reason: 'bad lesson', tier: 'hive' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(true);
-			expect(parsed.mode).toBe('archive');
-			expect(parsed.tier).toBe('hive');
-			expect(parsed.previous_status).toBe('established');
-			expect(parsed.status).toBe('archived');
-
-			const entries = await readKnowledge<HiveKnowledgeEntry>(hivePath);
-			expect(entries).toHaveLength(1);
-			expect(entries[0].status).toBe('archived');
-
-			// Hive tombstones go to the shared hive events log, not the
-			// project-local log.
-			const tomb = (await readHiveKnowledgeEvents()).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb).toHaveLength(1);
-			expect(tomb[0].entry_id).toBe('hive-1');
-			expect(tomb[0].tier).toBe('hive');
-			expect(tomb[0].previous_status).toBe('established');
-
-			// The project-local log must NOT receive the hive tombstone.
-			const localTomb = (await readKnowledgeEvents(dir)).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(localTomb).toHaveLength(0);
-		});
-
-		it('rejects quarantine when tier=hive and mode=quarantine (G5 #1716: swarm-only)', async () => {
-			const raw = await knowledge_archive.execute(
-				{
-					id: 'hive-1',
-					reason: 'suspect hive lesson',
-					tier: 'hive',
-					mode: 'quarantine',
-				},
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			// G5: hive-tier quarantine via the archive tool is rejected with a
-			// clear error. The old behavior silently flipped status in place
-			// and produced an unrestorable orphan; erroring is strictly better.
-			expect(parsed.success).toBe(false);
-			// PRR-011: pin the actionable guidance, not just the swarm-only fragment.
-			expect(parsed.error).toMatch(
-				/quarantine via the archive tool is swarm-only/i,
-			);
-			expect(parsed.error).toMatch(/\/swarm knowledge quarantine/);
-
-			// The hive entry is unchanged.
-			const entries = await readKnowledge<HiveKnowledgeEntry>(hivePath);
-			expect(entries).toHaveLength(1);
-			expect(entries[0].status).toBe('established');
-
-			// No tombstone written.
-			const tomb = (await readHiveKnowledgeEvents()).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb).toHaveLength(0);
-		});
-
-		it('purges hive entry with allow_purge:true when tier=hive', async () => {
-			const raw = await knowledge_archive.execute(
-				{
-					id: 'hive-1',
-					reason: 'purge bad hive',
-					tier: 'hive',
-					mode: 'purge',
-					allow_purge: true,
-				},
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(true);
-			expect(parsed.status).toBe('purged');
-			expect(parsed.tier).toBe('hive');
-
-			const entries = await readKnowledge<HiveKnowledgeEntry>(hivePath);
-			expect(entries).toHaveLength(0);
-
-			const tomb = (await readHiveKnowledgeEvents()).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(tomb[0].mode).toBe('purge');
-			expect(tomb[0].tier).toBe('hive');
-		});
-
-		it('refuses to purge hive entry without allow_purge:true', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'hive-1', reason: 'attempt purge', tier: 'hive', mode: 'purge' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(false);
-			expect(parsed.error).toContain('allow_purge');
-
-			const entries = await readKnowledge<HiveKnowledgeEntry>(hivePath);
-			expect(entries).toHaveLength(1);
-			expect(entries[0].status).toBe('established');
-		});
-
-		it('returns not found for unknown hive entry', async () => {
-			const raw = await knowledge_archive.execute(
-				{ id: 'unknown-hive', reason: 'does not exist', tier: 'hive' },
-				ctx(dir),
-			);
-			const parsed = JSON.parse(raw);
-			expect(parsed.success).toBe(false);
-			expect(parsed.message).toBe('entry not found');
-		});
-
-		it('records the hive tombstone in a shared log readable from any project', async () => {
-			// Archive a hive entry while running in "project A".
-			const projectA = join(dir, 'project-a');
-			mkdirSync(projectA, { recursive: true });
-			await knowledge_archive.execute(
-				{ id: 'hive-1', reason: 'remediated from project A', tier: 'hive' },
-				ctx(projectA),
-			);
-
-			// "Project B" — a different project directory on the same machine —
-			// reads the shared hive events log and sees the remediation.
-			const projectB = join(dir, 'project-b');
-			mkdirSync(projectB, { recursive: true });
-			const sharedTomb = (await readHiveKnowledgeEvents()).filter(
-				(e): e is ArchivedEvent => e.type === 'archived',
-			);
-			expect(sharedTomb).toHaveLength(1);
-			expect(sharedTomb[0].entry_id).toBe('hive-1');
-			expect(sharedTomb[0].reason).toBe('remediated from project A');
-
-			// Neither project's local event log holds the hive tombstone.
-			expect(await readKnowledgeEvents(projectA)).toHaveLength(0);
-			expect(await readKnowledgeEvents(projectB)).toHaveLength(0);
-		});
-	});
-});
-
-// ============================================================================
-// Post-archive hook — skill lifecycle integration (task 2.2)
-// ============================================================================
-
-describe('post-archive hook', () => {
-	let dir: string;
-	let origHome: string;
-
-	beforeEach(() => {
-		dir = join(
-			tmpdir(),
-			`ka-hook-${Date.now()}-${Math.random().toString(36).slice(2)}`,
-		);
-		mkdirSync(dir, { recursive: true });
-		origHome = process.env.HOME;
-		process.env.HOME = dir;
-	});
-
-	afterEach(() => {
-		process.env.HOME = origHome;
-		rmSync(dir, { force: true, recursive: true });
-	});
-
-	// -------------------------------------------------------------------------
-	// Context helper (mirrors existing test pattern)
-	// -------------------------------------------------------------------------
-	const ctx = (directory: string): any => ({
-		directory,
-		sessionID: 'sess-hook',
-		agent: 'architect',
-	});
-
-	// -------------------------------------------------------------------------
-	// Helpers
-	// -------------------------------------------------------------------------
-	async function appendKnowledgeEntry(
-		id: string,
-		status: string,
-	): Promise<void> {
-		const swarmPath = resolveSwarmKnowledgePath(dir);
-		await appendKnowledge(swarmPath, {
-			id,
-			tier: 'swarm',
-			lesson: `Lesson ${id}`,
-			category: 'process',
-			tags: [],
-			scope: 'global',
-			confidence: 0.5,
-			status,
-			confirmed_by: [],
-			project_name: 'test',
-			retrieval_outcomes: {
-				applied_count: 0,
-				succeeded_after_count: 0,
-				failed_after_count: 0,
-			},
-			schema_version: 2,
-			created_at: new Date().toISOString(),
-			updated_at: new Date().toISOString(),
-		} satisfies SwarmKnowledgeEntry);
-	}
-
-	// -------------------------------------------------------------------------
-	// Tests
-	// -------------------------------------------------------------------------
-
-	it('archives entry successfully (verifies post-archive hook is wired)', async () => {
-		await appendKnowledgeEntry('src-entry-1', 'candidate');
-
-		const raw = await knowledge_archive.execute(
-			{ id: 'src-entry-1', reason: 'test-reason' },
-			ctx(dir),
-		);
-		const result = JSON.parse(raw);
-
-		expect(result.success).toBe(true);
-		expect(result.id).toBe('src-entry-1');
-		expect(result.status).toBe('archived');
-	});
-
-	it('retires stale skill when all its sources are archived (file-based verification)', async () => {
-		// Write entry as 'candidate' so execute can archive it
-		await appendKnowledgeEntry('src-entry-1', 'candidate');
-
-		// Create a stale skill that depends on src-entry-1
-		const skillDir = join(
-			dir,
-			'.opencode',
-			'skills',
-			'generated',
-			'stale-skill-x',
-		);
-		mkdirSync(skillDir, { recursive: true });
-		writeFileSync(
-			join(skillDir, 'SKILL.md'),
-			[
-				'---',
-				'name: stale-skill-x',
-				'source_knowledge_ids:',
-				'  - src-entry-1',
-				'---',
-				'# Stale Skill X',
-			].join('\n'),
-		);
-		writeFileSync(join(skillDir, 'stale.marker'), 'needs regen\n');
-
-		const raw = await knowledge_archive.execute(
-			{ id: 'src-entry-1', reason: 'test-reason' },
-			ctx(dir),
-		);
-		const result = JSON.parse(raw);
-		expect(result.success).toBe(true);
-
-		// Wait for microtask to fire
-		await new Promise<void>((resolve) => setTimeout(resolve, 100));
-
-		// The microtask calls findStaleSkillsBySourceKnowledgeId (real, unpached) which
-		// finds our stale skill because it has sourceKnowledgeIds containing 'src-entry-1'.
-		// Then retireOrMarkStale is called with allArchivedIds={src-entry-1}.
-		// Since ALL sources are archived, retireSkill is called which creates retired.marker.
-		expect(existsSync(join(skillDir, 'retired.marker'))).toBe(true);
-	});
-
-	it('getArchivedKnowledgeIds returns archived and quarantined entry IDs', async () => {
-		const swarmPath = resolveSwarmKnowledgePath(dir);
-		await appendKnowledge(swarmPath, {
-			id: 'archived-entry',
-			tier: 'swarm',
-			lesson: 'L',
-			category: 'process',
-			tags: [],
-			scope: 'global',
-			confidence: 0.5,
-			status: 'archived',
-			confirmed_by: [],
-			project_name: 'test',
-			retrieval_outcomes: {
-				applied_count: 0,
-				succeeded_after_count: 0,
-				failed_after_count: 0,
-			},
-			schema_version: 2,
-			created_at: '',
-			updated_at: '',
-		} satisfies SwarmKnowledgeEntry);
-		await appendKnowledge(swarmPath, {
-			id: 'quarantined-entry',
-			tier: 'swarm',
-			lesson: 'L',
-			category: 'process',
-			tags: [],
-			scope: 'global',
-			confidence: 0.5,
-			status: 'quarantined',
-			confirmed_by: [],
-			project_name: 'test',
-			retrieval_outcomes: {
-				applied_count: 0,
-				succeeded_after_count: 0,
-				failed_after_count: 0,
-			},
-			schema_version: 2,
-			created_at: '',
-			updated_at: '',
-		} satisfies SwarmKnowledgeEntry);
-		await appendKnowledge(swarmPath, {
-			id: 'active-entry',
-			tier: 'swarm',
-			lesson: 'L',
-			category: 'process',
-			tags: [],
-			scope: 'global',
-			confidence: 0.5,
-			status: 'candidate',
-			confirmed_by: [],
-			project_name: 'test',
-			retrieval_outcomes: {
-				applied_count: 0,
-				succeeded_after_count: 0,
-				failed_after_count: 0,
-			},
-			schema_version: 2,
-			created_at: '',
-			updated_at: '',
-		} satisfies SwarmKnowledgeEntry);
-
-		const ids = await getArchivedKnowledgeIds(dir);
-		expect(ids).toContain('archived-entry');
-		expect(ids).toContain('quarantined-entry');
-		expect(ids).not.toContain('active-entry');
-	});
-});
+/**
+ * knowledge-archive.test.ts — barrel re-export
+ *
+ * Part of F-003 fix: split into focused sub-files to comply with
+ * AGENTS.md invariant 7 / FR-006 500-line cap.
+ *
+ * Original 646-line file split into:
+ * - knowledge-archive.tool-swarm.test.ts  — swarm-tier archive tests
+ * - knowledge-archive.tool-hive.test.ts   — hive-tier archive tests
+ * - knowledge-archive.tool-post-archive-hook.test.ts — post-archive hook tests
+ */
+import './knowledge-archive.tool-swarm.test';
+import './knowledge-archive.tool-hive.test';
+import './knowledge-archive.tool-post-archive-hook.test';
